cleanup: move base monitoring resources to alarms
